% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculate Mean and Standard Deviation}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list containing the mean and standard deviation.
}
\description{
Calculates the mean and standard deviation of a numeric vector.
}
\examples{
mean_sd(c(1, 2, 3, 4, 5))
}
