% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_string.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse a String}
\usage{
reverse_string(str)
}
\arguments{
\item{str}{A character string.}
}
\value{
A character string with reversed characters.
}
\description{
Reverses the characters in a given string.
}
\examples{
reverse_string("Hello")
}
