% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-test.R
\name{t_test_example}
\alias{t_test_example}
\title{Perform T-Test}
\usage{
t_test_example(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The result of the t-test.
}
\description{
Performs a two-sample t-test on two numeric vectors.
}
\examples{
t_test_example(c(1, 2, 3), c(4, 5, 6))
}
